fix 64bit

+++ base/Condition.h
@@ -33,12 +33,12 @@ public:
 	}
 
 	/* timedwait using
-	 * a microseconds time value.*/
-	bool timedwait(uint32_t timeout);
+	 * a nanoseconds time value.*/
+	bool timedwait(uint64_t timeout);
 
 private:
 	static const uint32_t kNanoSecondsPerSecond = 1e9;
-
+	
 	Mutex& mutex_;
 	pthread_cond_t cond_;
 };
